XIVY-2942 refactor: improve folder structure
