+++ packages/dataclass-editor/src/DataClassEditor.tsx
@@ -4,8 +4,8 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useEffect, useMemo, useState } from 'react';
 import type { DataClass } from './components/dataclass/data/dataclass';
 import { headerTitles } from './components/dataclass/data/dataclass-utils';
-import { DataClassDetailContent } from './components/dataclass/detail/DataClassDetailContent';
-import { FieldDetailContent } from './components/dataclass/detail/FieldDetailContent';
+import { DataClassDetailContent } from './components/dataclass/detail/dataclass/DataClassDetailContent';
+import { FieldDetailContent } from './components/dataclass/detail/field/FieldDetailContent';
 import { DataClassMasterContent } from './components/dataclass/master/DataClassMasterContent';
 import { DataClassMasterToolbar } from './components/dataclass/master/DataClassMasterToolbar';
 import { AppProvider } from './context/AppContext';
