+++ packages/dataclass-editor/src/components/dataclass/detail/field/FieldDetailContent.css
